+++ tests/managed_deadlock.rs
@@ -8,7 +8,7 @@ use tokio::{
     task, time,
 };
 
-use deadpool::managed::{self, RecycleError, RecycleResult};
+use deadpool::managed::{self, Metrics, RecycleError, RecycleResult};
 
 type Pool = managed::Pool<Manager>;
 
@@ -66,7 +66,7 @@ impl managed::Manager for Manager {
         self.create_rx.lock().await.recv().await.unwrap()
     }
 
-    async fn recycle(&self, _conn: &mut ()) -> RecycleResult<()> {
+    async fn recycle(&self, _conn: &mut (), _: &Metrics) -> RecycleResult<()> {
         match self.recycle_rx.lock().await.recv().await.unwrap() {
             Ok(()) => Ok(()),
             Err(e) => Err(RecycleError::Backend(e)),
@@ -87,20 +87,23 @@ async fn pool_drained() {
     let get_1 = tokio::spawn(async move { pool_clone.get().await });
     task::yield_now().await;
     assert_eq!(pool.status().size, 0);
-    assert_eq!(pool.status().available, -1);
+    assert_eq!(pool.status().available, 0);
+    assert_eq!(pool.status().waiting, 1);
 
     // let second task wait for the connection
     let pool_clone = pool.clone();
     let get_2 = tokio::spawn(async move { pool_clone.get().await });
     task::yield_now().await;
     assert_eq!(pool.status().size, 0);
-    assert_eq!(pool.status().available, -2);
+    assert_eq!(pool.status().available, 0);
+    assert_eq!(pool.status().waiting, 2);
 
     // first task receives an error
     rc.create_err();
     assert!(get_1.await.unwrap().is_err());
     assert_eq!(pool.status().size, 0);
-    assert_eq!(pool.status().available, -1);
+    assert_eq!(pool.status().available, 0);
+    assert_eq!(pool.status().waiting, 1);
 
     // the second task should now be able to create an object
     rc.create_ok();
@@ -108,10 +111,12 @@ async fn pool_drained() {
     assert!(get_2_result.is_ok(), "get_2 should not time out");
     assert_eq!(pool.status().size, 1);
     assert_eq!(pool.status().available, 0);
+    assert_eq!(pool.status().waiting, 0);
     assert!(
         get_2_result.unwrap().unwrap().is_ok(),
         "get_2 should receive an object"
     );
     assert_eq!(pool.status().size, 1);
     assert_eq!(pool.status().available, 1);
+    assert_eq!(pool.status().waiting, 0);
 }