+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## v0.10.0 (unreleased)
+
+- Remove unreachable enum variant `BuildError::Backend`
+- Split `Status.available` into `available` and `waiting`.
+- Add `QueueMode` configuration option for choosing between
+  a `FIFO` (default) and `LIFO` queue.
+- Remove `HookError::Continue` and `HookError::Abort` variants
+  replacing it with the contents of `HookErrorCause`. Returning
+  a `HookError` from a `post_create` hook causes the `Pool::get`
+  operation to fail while returning it from a `pre_recycle` or
+  `post_recycle` hook the operation continues.
+- Add `metrics` argument to `Manager::recycle` method.
+
 ## v0.9.5
 
 - Fix bug causing the pool to exceed its `max_size` in the