+++ redis-cluster/CHANGELOG.md
@@ -1,5 +1,5 @@
 # Change Log
 
-## v0.1.0
+## v0.1.0 (unreleased)
 
 * First release