+++ sqlite/CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.6.0 (unreleased)
+
+* Update `deadpool` dependency to version `0.10`
+* Update `rusqlite` dependency to version `0.29`
+
 ## v0.5.0
 
 - Update `rusqlite` dependency to version `0.28`